requirements.txt
ollama>=0.2.0
chromadb>=0.5.0
loguru>=0.7.0
psutil>=5.9.0
httpx>=0.27.0
apscheduler>=3.10.0
pydantic>=2.0.0
tenacity>=8.2.0
rich>=13.0.0
langchain-core>=0.2.0
pytest>=8.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
websockets>=12.0
python-multipart>=0.0.9
aiofiles>=23.0.0
python-dotenv>=1.0.0
